tests/wpt/web-platform-tests/cookies/prefix/__host.document-cookie.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/cookies/resources/cookie-helper.sub.js"></script>
<script>
  ["", "MaxAge=10"].forEach(extraParams => {
    // Without 'secure'
    set_prefixed_cookie_via_dom_test({
      prefix: "__Host-",
      params: "Path=/;" + extraParams,
      shouldExistInDOM: false,
      shouldExistViaHTTP: false,
      title: "__Host: Secure origin: Does not set 'Path=/;" + extraParams + "'"
    });

    set_prefixed_cookie_via_dom_test({
      prefix: "__HoSt-",
      params: "Path=/;" + extraParams,
      shouldExistInDOM: false,
      shouldExistViaHTTP: false,
      title: "__HoSt: Secure origin: Does not set 'Path=/;" + extraParams + "'"
    });

    // With 'secure'
    set_prefixed_cookie_via_dom_test({
      prefix: "__Host-",
      params: "Secure; Path=/;" + extraParams,
      shouldExistInDOM: true,
      shouldExistViaHTTP: true,
      title: "__Host: Secure origin: Does set 'Secure; Path=/;" + extraParams + "'"
    });

    set_prefixed_cookie_via_dom_test({
      prefix: "__HoSt-",
      params: "Secure; Path=/;" + extraParams,
      shouldExistInDOM: true,
      shouldExistViaHTTP: true,
      title: "__HoSt: Secure origin: Does set 'Secure; Path=/;" + extraParams + "'"
    });

    // With 'domain'
    set_prefixed_cookie_via_dom_test({
      prefix: "__Host-",
      params: "Secure; Path=/; Domain=" + document.location.hostname + "; " + extraParams,
      shouldExistInDOM: false,
      shouldExistViaHTTP: false,
      title: "__Host: Secure origin: Does not set 'Secure; Path=/; Domain=" + document.location.hostname + "; " + extraParams + "'"
    });

    set_prefixed_cookie_via_dom_test({
      prefix: "__HoSt-",
      params: "Secure; Path=/; Domain=" + document.location.hostname + "; " + extraParams,
      shouldExistInDOM: false,
      shouldExistViaHTTP: false,
      title: "__HoSt: Secure origin: Does not set 'Secure; Path=/; Domain=" + document.location.hostname + "; " + extraParams + "'"
    });
  });

  set_prefixed_cookie_via_dom_test({
    prefix: "__Host-",
    params: "Secure; Path=/cookies/resources/list.py",
    shouldExistInDOM: false,
    shouldExistViaHTTP: false,
    title: "__Host: Secure origin: Does not set 'Secure; Path=/cookies/resources/list.py'"
  });

  set_prefixed_cookie_via_dom_test({
    prefix: "__HoSt-",
    params: "Secure; Path=/cookies/resources/list.py",
    shouldExistInDOM: false,
    shouldExistViaHTTP: false,
    title: "__HoSt: Secure origin: Does not set 'Secure; Path=/cookies/resources/list.py'"
  });
</script>
