LayoutTests/imported/w3c/web-platform-tests/cookies/prefix/__host.document-cookie.https-expected.txt


PASS __Host: Secure origin: Does not set 'Path=/;'
FAIL __HoSt: Secure origin: Does not set 'Path=/;' assert_equals: `__HoSt-prefixtestcookie=0.5989997277319348` in `document.cookie` expected false but got true
PASS __Host: Secure origin: Does set 'Secure; Path=/;'
PASS __HoSt: Secure origin: Does set 'Secure; Path=/;'
PASS __Host: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; '
FAIL __HoSt: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; ' assert_equals: `__HoSt-prefixtestcookie=0.4202761366018031` in `document.cookie` expected false but got true
PASS __Host: Secure origin: Does not set 'Path=/;MaxAge=10'
FAIL __HoSt: Secure origin: Does not set 'Path=/;MaxAge=10' assert_equals: `__HoSt-prefixtestcookie=0.21196434592296443` in `document.cookie` expected false but got true
PASS __Host: Secure origin: Does set 'Secure; Path=/;MaxAge=10'
PASS __HoSt: Secure origin: Does set 'Secure; Path=/;MaxAge=10'
PASS __Host: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; MaxAge=10'
FAIL __HoSt: Secure origin: Does not set 'Secure; Path=/; Domain=web-platform.test; MaxAge=10' assert_equals: `__HoSt-prefixtestcookie=0.09791975455374136` in `document.cookie` expected false but got true
PASS __Host: Secure origin: Does not set 'Secure; Path=/cookies/resources/list.py'
FAIL __HoSt: Secure origin: Does not set 'Secure; Path=/cookies/resources/list.py' assert_equals: expected (undefined) undefined but got (string) "0.5374139193588056"
